Changes while switching the Python exposure to ndobject

+++ include/dynd/shape_tools.hpp
@@ -26,8 +26,8 @@ bool shape_can_broadcast(int dst_ndim, const intptr_t *dst_shape,
 inline bool shape_can_broadcast(const std::vector<intptr_t>& dst_shape,
                         const std::vector<intptr_t>& src_shape)
 {
-    return shape_can_broadcast(dst_shape.size(), dst_shape.empty() ? NULL : &dst_shape[0],
-                        src_shape.size(), src_shape.empty() ? NULL : &src_shape[0]);
+    return shape_can_broadcast((int)dst_shape.size(), dst_shape.empty() ? NULL : &dst_shape[0],
+                        (int)src_shape.size(), src_shape.empty() ? NULL : &src_shape[0]);
 }
 
 /**
@@ -118,7 +118,7 @@ inline void multistrides_to_axis_perm(int ndim, int noperands, intptr_t **operst
 void print_shape(std::ostream& o, int ndim, const intptr_t *shape);
 
 inline void print_shape(std::ostream& o, const std::vector<intptr_t>& shape) {
-    print_shape(o, shape.size(), shape.empty() ? NULL : &shape[0]);
+    print_shape(o, (int)shape.size(), shape.empty() ? NULL : &shape[0]);
 }
 
 /**
